✏️ 전반적인 오류 수정
